add button to add new server (just design)
