+++ web/src/css/mode.less
@@ -80,4 +80,18 @@
     .mode-reverse-dropdown {
         margin: 0 5px;
     }
+
+    .mode-reverse-add-server {
+        margin-left: 10px;
+        margin-top: 15px;
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: center;
+        
+        i {
+            font-size: 2rem;
+            margin-right: 5px;
+        }
+    }
 }
